MAINT: Restructuring of src
