+++ src/mdestimates/error_evaluation/diffusive_error.py
